Desktop/mental-health-app/ml-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
torch==2.1.0
torch-geometric==2.4.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
pydantic==2.4.2
python-multipart==0.0.6
joblib==1.3.2
networkx==3.1
matplotlib==3.7.2
seaborn==0.12.2
pymongo==4.5.0
python-dotenv==1.0.0
requests==2.31.0